# Reduce TTF`@mtkmodel`