--- a/src/ModelingToolkit.jl
+++ b/src/ModelingToolkit.jl
@@ -225,11 +225,20 @@ PrecompileTools.@compile_workload begin
     @named sys = ODESystem([ModelingToolkit.D_nounits(x) ~ -x], ModelingToolkit.t_nounits)
     prob = ODEProblem(structural_simplify(sys), [x => 30.0], (0, 100), [], jac = true)
     @mtkmodel __testmod__ begin
+        @constants begin
+            c = 1.0
+        end
         @structural_parameters begin
             structp = false
         end
-        @variables begin
-            x(t) = 0.0, [description="foo", guess=1.0]
+        if structp
+            @variables begin
+                x(t) = 0.0, [description="foo", guess=1.0]
+            end
+        else
+            @variables begin
+                x(t) = 0.0, [description="foo w/o structp", guess=1.0]
+            end
         end
         @parameters begin
             a = 1.0, [description="bar"]