This PR attemts to reduce invaldiations of `@mtkmodel` code and thus enables precompilation of said macro. 

Still WIP, so far I've only removed invalidations on the toplevel `_model_macro` function, but this allready leads to a ~10s improvement in TTFX.

Below examples executed in new julia session
```julia
using ModelingToolkit
@time @eval @mtkmodel Mod begin
end;
# this PR
# 0.185190 seconds (735.61 k allocations: 37.985 MiB, 16.39% gc time, 97.47% compilation time: 91% of which was recompilation)
# master
# 16.592960 seconds (114.90 M allocations: 5.945 GiB, 9.04% gc time, 99.97% compilation time)
```
```julia
using ModelingToolkit
@time @eval @mtkmodel Mod begin
    @variables begin
        x(t)
        y(t)
    end
    @parameters begin
        p
    end
    @equations begin
        x ~ y + p
    end
end;
# this PR
# 12.499833 seconds (82.14 M allocations: 4.328 GiB, 9.96% gc time, 99.93% compilation time: 96% of which was recompilation)
# master
# 22.175370 seconds (121.23 M allocations: 6.330 GiB, 5.06% gc time, 99.96% compilation time: <1% of which was recompilation)
```

## Checklist

- [ ] Appropriate tests were added
- [ ] Any code changes were done in a way that does not break public API
- [ ] All documentation related to code changes were updated
- [ ] The new code follows the
  [contributor guidelines](https://github.com/SciML/.github/blob/master/CONTRIBUTING.md), in particular the [SciML Style Guide](https://github.com/SciML/SciMLStyle) and
  [COLPRAC](https://github.com/SciML/COLPRAC).
- [ ] Any new documentation only uses public API
  
## Additional context

Add any other context about the problem here.