@hexaeder 2025-03-19
Hm in the end it seems all it takes is full despecialization of two recursive parser function which are very hard for the compiler.

My initial goal was to fix the invalidations (i.e. no "recompilation" on first call), but I didn't manage. However, the despecialization + precomile workload alone leads to quite the speedup...
```
restart()
using ModelingToolkit
@time @eval @mtkmodel Mod begin
end;
# this PR
# 0.851972 seconds (3.93 M allocations: 202.397 MiB, 4.32% gc time, 99.68% compilation time: 88% of which was recompilation)
# master
# 16.592960 seconds (114.90 M allocations: 5.945 GiB, 9.04% gc time, 99.97% compilation time)

restart()
using ModelingToolkit
@time @eval @mtkmodel Mod begin
    @variables begin
        x(t)
        y(t)
    end
    @parameters begin
        p
    end
    @equations begin
        x ~ y + p
    end
end;
# this PR
# 3.384142 seconds (5.87 M allocations: 305.992 MiB, 1.34% gc time, 99.75% compilation time: 41% of which was recompilation)
# master
# 22.175370 seconds (121.23 M allocations: 6.330 GiB, 5.06% gc time, 99.96% compilation time: <1% of which was recompilation)

restart(false)
@time using ModelingToolkit
@time @eval @mtkmodel Mod begin
    @constants begin
        c = 1.0
    end
    @structural_parameters begin
        structp = false
    end
    @variables begin
        x(t) = 0.0, [description="foo", guess=1.0]
    end
    @parameters begin
        a = 1.0, [description="bar"]
        if structp
            b=2*a, [description="if"]
        else
            c
        end
    end
    @equations begin
        x ~ a + b
    end
end;
# this PR
# 4.351150 seconds (7.32 M allocations: 378.812 MiB, 1.70% gc time, 99.69% compilation time: 47% of which was recompilation)
# master
# 24.336051 seconds (125.52 M allocations: 6.563 GiB, 6.53% gc time, 99.95% compilation time: <1% of which was recompilation)
```