@ChrisRackauckas 2025-03-20
Super awesome, thanks for digging into this!